feat: clean napi-derive noop feature code path
