+++ crates/macro/src/expand.rs
@@ -1,0 +1,9 @@
+cfg_if::cfg_if! {
+  if #[cfg(feature = "noop")] {
+    mod noop;
+    pub use self::noop::*;
+  } else {
+    mod napi;
+    pub use self::napi::*;
+  }
+}
